[dev]project refactored
