+++ flaskivr/__init__.py
@@ -1,0 +1,10 @@
+from flask import Flask
+from .routes.ivr_url import IVRFlow  
+from .routes.auth import Auth
+
+app = Flask(__name__)
+
+app.register_blueprint(IVRFlow)
+app.register_blueprint(Auth,url_prefix='/authenticate')
+
+
