+++ flaskivr/settings.py
@@ -1,0 +1,13 @@
+import os
+
+# cred_json = os.environ['json_path']
+# lst_num = ['first', 'second', 'third', 'forth', 'fifth', 'sixth', 'seventh', 'eighth', 'ninth', 'tenth']
+# main_number = os.environ['main_number']
+# optional_number = os.environ['optional_number']
+# postgreSQLpass = os.environ['postgreSQLpass']
+
+
+# #import pprint
+
+# my_api_key = os.environ['google_api_key']
+# my_cse_id = os.environ['google_cse_id']
