+++ flaskivr/tools/templates/handshake.png
